refactor: move to schema
